Updated level system and added level changing
I completely overhauled the level system, adding the level class and a
class for each level we have so far (mine and joes). I also made the
levelEnd sprite that can be added to any level to mark the end and will
cause the next level (or screen) to be called.