IT'S WORKING! IT'S WORKING
So, i got the movement to work with the buttons. Key functions to look
at are goodPress() and inputAct() in playscreen. Also, there are still
some issues with the placement of the buttons (they're a bit high) but
if i lower them it fucks shit up and i dont have the time to fix it
right now. Anyways, working hud and movement is now at Beta level.