Camera and Roadblock adjustments
First off, im sorry if i cause any clonflicts with this, i had to change
some peoples stuff.... However, the camera is now a bit better, still
some fixing that needs done, but it at least shows enough of the screen
that the game is playable now. Also, i revamped RoadBlock to just check
circle vs rectangle objects, so if you make a roadblock it can be
either. if you make it an elipse it will just draw it as a circle, so
it's best to just use circles unless it doesnt really matter. Anywhoo,
next up....fixing buttons.....